+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # NEXT RELEASE
 * Remove `extract_pages_from_pdf` command line tool (moved to `pdfalyzer`)
+* Use `FreeSimpleGUI` instead of `PySimpleGUI`
 
 ### 1.13.4
 * Upgrade `PyPDF` to 6.0.0